// src/app/api/model/contributionContributionCreateReq.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ContributionArticleReq } from './contributionArticleReq';
import { ContributionImageCreateReq } from './contributionImageCreateReq';


export interface ContributionContributionCreateReq { 
    article?: ContributionArticleReq;
    images?: Array<ContributionImageCreateReq>;
}