// src/app/api/model/contribution-contribution-create-req.ts
/**
 * 123
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ContributionImageCreateReq } from './contribution-image-create-req';
import { ContributionArticleReq } from './contribution-article-req';


export interface ContributionContributionCreateReq { 
    article?: ContributionArticleReq;
    description?: string;
    images?: Array<ContributionImageCreateReq>;
    title?: string;
}